+++ mne/decoding/gsoc.py
@@ -5,7 +5,7 @@ from scipy import linalg
 
 from .. import Covariance
 from ..cov import _regularized_covariance
-from ..preprocessing.xdawn import Xdawn
+from ..preprocessing import Xdawn
 
 
 class XdawnTransformer(Xdawn):
@@ -16,15 +16,17 @@ class XdawnTransformer(Xdawn):
     to signal + noise ratio (SSNR) of the ERP responses. Xdawn was originally
     designed for P300 evoked potential by enhancing the target response with
     respect to the non-target response. This implementation is a light version
-    of the original version. This does not handle overlap cases however it is
-    helpful while pipelining.
+    of the original version and follows scikit-learn API strictly.
+
+    .. note:: This does not handle overlapping events. Use original
+              preprocessing.xdawn instead.
 
     Parameters
     ----------
-    n_chan : int
-         Integer indicating the number of channels.
     n_components : int (default 2)
         The number of components to decompose M/EEG signals.
+    n_chan : int
+         Integer indicating the number of channels.
     signal_cov : None | Covariance | ndarray, shape (n_channels, n_channels)
         (default None). The signal covariance used for whitening of the data.
         if None, the covariance is estimated from the epochs signal.
@@ -33,10 +35,6 @@ class XdawnTransformer(Xdawn):
         if float, shrinkage covariance is used (0 <= shrinkage <= 1).
         if str, optimal shrinkage using Ledoit-Wolf Shrinkage ('ledoit_wolf')
         or Oracle Approximating Shrinkage ('oas').
-    correct_overlap : bool (default False)
-        Apply correction for overlaped ERP for the estimation of evokeds
-        responses. if 'auto', the overlapp correction is chosen in function
-        of the events in epochs.events.
 
     Attributes
     ----------
@@ -46,12 +44,11 @@ class XdawnTransformer(Xdawn):
     patterns_ : dict of ndarray
         If fit, the Xdawn patterns used to restore M/EEG signals for each event
         type, else empty.
-    event_id_mean_ : dict of np.ndarray
-        If fit, the mean of each event type.
+    
 
     See Also
     --------
-    ICA
+    mne.preprocessing.Xdawn
     CSP
 
     References
@@ -66,19 +63,16 @@ class XdawnTransformer(Xdawn):
     2011 19th European (pp. 1382-1386). IEEE.
     """
 
-    def __init__(self, n_chan, n_components=2, signal_cov=None,
-                 reg=None, correct_overlap=False):
+    def __init__(self, n_components=2, n_chan=None, signal_cov=None,
+                 reg=None):
         """init xdawn."""
+        if n_chan is None:
+            raise ValueError("n_chan cannot be none. Please provide the "
+                             "number of channels")
         self.n_chan = n_chan
         self.n_components = n_components
         self.signal_cov = signal_cov
         self.reg = reg
-        if correct_overlap:
-            raise DeprecationWarning("correct_overlap feature is deprecated "
-                                     "temporarily.")
-        self.filters_ = dict()
-        self.patterns_ = dict()
-        self.event_id_mean_ = dict()
 
     def fit(self, X, y):
         """Fit Xdawn from epochs.
@@ -86,18 +80,18 @@ class XdawnTransformer(Xdawn):
         Parameters
         ----------
         X : ndarray, shape(n_channels, n_times * n_freq)
-            Data of epochs
+            Data of epochs.
         y : ndarray shape(n_samples,)
-            Target values
+            Target values.
 
         Returns
         -------
-        self : Xdawn instance
-            The Xdawn instance.
+        self : XdawnTransformer instance
+            The XdawnTransformer instance.
         """
         from sklearn.preprocessing import LabelEncoder
 
-        if X.ndim is not 2 or not isinstance(X, np.ndarray):
+        if X.ndim != 2 or not isinstance(X, np.ndarray):
             raise ValueError("X should be 2 dimensional ndarray")
         if not isinstance(y, np.ndarray):
             raise ValueError("Labels must be numpy array")
@@ -117,7 +111,9 @@ class XdawnTransformer(Xdawn):
                              'or a ndarray')
 
         # estimates evoked covariance
-        self.event_id_cov_ = dict()
+        self.filters_ = dict()
+        self.patterns_ = dict()
+        event_id_cov_ = dict()
         event_id_mean = dict()
         toeplitz = dict()
         classes = LabelEncoder().fit(y).classes_
@@ -127,20 +123,19 @@ class XdawnTransformer(Xdawn):
                                 shape[2]), axis=0)
             event_id_mean[eid] = mean_data
             toeplitz[eid] = 1.0
-        self.event_id_mean_ = event_id_mean
 
         for eid in classes:
             data = np.dot(event_id_mean[eid], toeplitz[eid])
-            self.event_id_cov_[eid] = _regularized_covariance(data, self.reg)
+            event_id_cov_[eid] = _regularized_covariance(data, self.reg)
 
         # estimates spatial filters
         for eid in classes:
 
-            if self.signal_cov_.shape != self.event_id_cov_[eid].shape:
+            if self.signal_cov_.shape != event_id_cov_[eid].shape:
                 raise ValueError('Size of signal cov must be the same as the'
-                                 ' number of channels in epochs')
+                                 ' number of channels in the epochs')
 
-            evals, evecs = linalg.eigh(self.event_id_cov_[eid],
+            evals, evecs = linalg.eigh(event_id_cov_[eid],
                                        self.signal_cov_)
             evecs = evecs[:, np.argsort(evals)[::-1]]  # sort eigenvectors
             evecs /= np.sqrt(np.sum(evecs ** 2, axis=0))
@@ -159,7 +154,7 @@ class XdawnTransformer(Xdawn):
         Parameters
         ----------
         X : ndarray, shape(n_channels, n_times * n_freq)
-            data of epochs
+            data of epochs.
 
         Returns
         -------
@@ -178,7 +173,7 @@ class XdawnTransformer(Xdawn):
         # create full matrix of spatial filter
         full_filters = list()
         for filt in self.filters_.values():
-            full_filters.append(filt[:, 0:self.n_components])
+            full_filters.append(filt[:, :self.n_components])
         full_filters = np.concatenate(full_filters, axis=1)
 
         # Apply spatial filters
@@ -193,9 +188,9 @@ class XdawnTransformer(Xdawn):
         Parameters
         ----------
         X : ndarray, shape(n_channels, n_times * n_freq)
-            data of epochs
+            data of epochs.
         y : ndarray, shape(n_samples,)
-            labels of data
+            labels of data.
 
         Returns
         -------
@@ -216,6 +211,18 @@ class XdawnTransformer(Xdawn):
         Parameters
         ----------
         X : np.ndarray, shape(n_epochs, n_components * event_types * n_times)
+            The signal data to undergo inverse transform.
+        event_id : dict | list of str | None (default None)
+            The kind of event to apply. if None, a dict of inst will be return
+            one for each type of event xdawn has been fitted.
+        include : array_like of int | None (default None)
+            The indices referring to columns in the ummixing matrix. The
+            components to be kept. If None, the first n_components (as defined
+            in the Xdawn constructor) will be kept.
+        exclude : array_like of int | None (default None)
+            The indices referring to columns in the ummixing matrix. The
+            components to be zeroed out. If None, all the components except the
+            first n_components will be exclude.
 
         Returns
         -------
@@ -243,7 +250,30 @@ class XdawnTransformer(Xdawn):
         return data_dict
 
     def apply(self, inst, event_id=None, include=None, exclude=None):
-        """apply in this version of xdawn is not supported"""
+        """apply in this version of xdawn is not supported.
+
+        Parameters
+        ----------
+        inst : instance of Raw | Epochs | Evoked
+            The data to be processed.
+        event_id : dict | list of str | None (default None)
+            The kind of event to apply. if None, a dict of inst will be return
+            one for each type of event xdawn has been fitted.
+        include : array_like of int | None (default None)
+            The indices referring to columns in the ummixing matrix. The
+            components to be kept. If None, the first n_components (as defined
+            in the Xdawn constructor) will be kept.
+        exclude : array_like of int | None (default None)
+            The indices referring to columns in the ummixing matrix. The
+            components to be zeroed out. If None, all the components except the
+            first n_components will be exclude.
+
+        Returns
+        -------
+        None
+        """
 
         warnings.warn("Warning, apply is not supported in this version of "
-                      "xdawn. Use inverse_transform method instead")
+                      "xdawn. Use inverse_transform method instead.")
+
+        pass